Refactor: Move message receiving code into seperate class
